nlp_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
transformers==4.35.2
python-dotenv==1.0.0
redis==5.0.1